python/pyforge/requirements.txt
torch
torchvision
transformers
diffusers
llama-cpp-python
fastapi
uvicorn
ttkbootstrap
huggingface_hub
pillow
accelerate
audiocraft
TTS
openai-whisper
imageio
numpy
python-multipart
